Missing in_place_factory header and cmake fix
bonjour announcer.cpp and monitor.cpp are missing the header
boost/utility/in_place_factory.hpp.
In FindBonjour the mark_as_advanced function is moved to the end of
the file. If it is at the top it will clear Bonjour_LIBRARY and
Bonjour_INCLUDE_DIR arguments which are meant for to hint find_file
and find_library of locations. See the following link for more info
https://gitlab.kitware.com/cmake/cmake/issues/15448.